backend/requirements.txt
fastapi
uvicorn
redis
celery
py-clob-client
duckduckgo-search
langchain-community
google-genai
httpx
pydantic-settings
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
qdrant-client
fastembed
web3
